+++ docs/fr_FR/changelog.md
@@ -2,6 +2,7 @@
 
 - Support du ping ARP (nécessite de forcer l'installation des dépendances)
 - Support du ping sur un port (socket)
+- Amélioration du panel
 
 # 16/04/2018